// agentic-server/generated/prisma/models/State.ts
export type StateModel = runtime.Types.Result.DefaultSelection<Prisma.$StatePayload>
// ...
export type $StatePayload<ExtArgs extends runtime.Types.Extensions.InternalArgs = runtime.Types.Extensions.DefaultArgs> = {
  name: "State"
  objects: {
    query: Prisma.$queriesPayload<ExtArgs>
  }
  scalars: runtime.Types.Extensions.GetPayloadResult<{
    id: number
    response_id: string
    query_id: number
    input_tokens: number
    total_output_tokens: number
    reasoning_output_tokens: number
    tokens_per_second: number
    time_to_first_token_seconds: number
    createdAt: Date | null
  }, ExtArgs["result"]["state"]>
  composites: {}
}